Change primary key type from Long to Integer in repository and service classes

+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Controller/CategoriaInventarioControlador.java
@@ -38,7 +38,7 @@ public class CategoriaInventarioControlador {
 
     @DeleteMapping("/{userId}/{categoryId}")
     public ResponseEntity<String> borrarCategoria(@PathVariable Integer userId
-                                                , @PathVariable Long categoryId){
+                                                , @PathVariable Integer categoryId){
         Optional<CategoriaInventario> categoria = servicio.findById(categoryId);
         if(categoria.isPresent()){
             auditoria.guardarAccion(userId, "Categoria borrada", "categoria_inventario");
