+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Repository/ProveedorRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.uia.ing.soft.olda.dunamys.ing_software_dunamys.Model.Proveedor;
 
 @Repository
-public interface ProveedorRepositorio extends JpaRepository<Proveedor, Long>{
+public interface ProveedorRepositorio extends JpaRepository<Proveedor, Integer>{
 
 }
